gRPC contract refactoring + mapper usage improvements (#14)
* Refactored data fetching and removed unnecessary mappers.

* Refactor player retrieval for improved efficiency.

* Refactor AvroMapper and add logging to PlayerApi.

* Enhance player task generation to include level data.

* Updates solo-leveling-avro dependency to version 2.1.3.

* Refactored currency reward type to Int.

* Updated proto dependency and mapper.

* Refined logging and added null check strategy.

* Refactored task status handling with new service.

* Refactored player task status updates.

+++ solo-leveling-player-model/src/main/kotlin/com/sleepkqq/sololeveling/player/model/entity/task/Task.kt
@@ -6,7 +6,6 @@ import com.sleepkqq.sololeveling.player.model.entity.task.enums.TaskRarity
 import com.sleepkqq.sololeveling.player.model.entity.task.enums.TaskTopic
 import org.babyfish.jimmer.sql.*
 import org.babyfish.jimmer.sql.meta.UUIDIdGenerator
-import java.math.BigDecimal
 import java.util.UUID
 
 @Entity
@@ -24,7 +23,7 @@ interface Task : Model {
 
 	val experience: Int?
 
-	val currencyReward: BigDecimal?
+	val currencyReward: Int?
 
 	val rarity: TaskRarity?
